testing for gameBoard and Player
