+++ index.html
@@ -137,7 +137,7 @@
     integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
     crossorigin="anonymous"></script>
 
-  <script src="src/app.js"></script>
+  <script src="dist/main.js"></script>
 </body>
 
 </html>
