+++ src/gameBoard.js
@@ -1,0 +1,12 @@
+const gameBoard = (() => {
+  const renderBoard = () => {
+    let cell = '';
+    for (let i = 1; i <= 9; i += 1) {
+      cell += `<button class="cell cell-${i}" id="cell-${i}"></button>`;
+    }
+    document.getElementById('display-board').innerHTML = cell;
+  };
+  return { renderBoard };
+})();
+
+export default gameBoard;
